requirements.txt
groq
requests
python-dotenv
rich
flask
hindsight-client